+++ src/main/java/org/primal/mixin/EntityTypeMixin.java
@@ -1,5 +1,6 @@
 package org.primal.mixin;
 
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.entity.EntityDimensions;
 import net.minecraft.world.entity.EntityType;
 import org.primal.Primal_Main;
@@ -18,7 +19,7 @@ public class EntityTypeMixin {
             at = @At("HEAD"), cancellable = true)
     private void primal$modifyHitboxesForNewModels(CallbackInfoReturnable<EntityDimensions> cir) {
 
-        // Fox Hitbox: 0.6 x 0.7 -> 0.75 x 0.85
+        // Fox Hitbox: 0.6 x 0.7 -> 0.6 x 0.85
         if(p$THIS == EntityType.FOX && Primal_Main.ConfigCache.foxModelChange){
             cir.setReturnValue(EntityDimensions.scalable(0.6f, 0.85F));
         }
@@ -34,8 +35,16 @@ public class EntityTypeMixin {
         }
 
         // Wolf Hitbox: 0.6 x 0.5 -> 1.0 x 1.15
-        if(p$THIS == EntityType.WOLF && Primal_Main.ConfigCache.wolfModelChange){
-            cir.setReturnValue(EntityDimensions.scalable(1.0F, 1.15F));
+        if(Primal_Main.ConfigCache.wolfModelChange){
+            if(p$THIS == EntityType.WOLF)
+                cir.setReturnValue(EntityDimensions.scalable(1.0F, 1.15F));
+
+            //Pet cemetery compat
+            if(BuiltInRegistries.ENTITY_TYPE.getKey(p$THIS).getNamespace().equals("pet_cemetery")
+                    && (BuiltInRegistries.ENTITY_TYPE.getKey(p$THIS).getPath().equals("zombie_wolf") ||
+                        BuiltInRegistries.ENTITY_TYPE.getKey(p$THIS).getPath().equals("skeleton_wolf"))){
+                cir.setReturnValue(EntityDimensions.scalable(1.0F, 1.15F));
+            }
         }
 
         // Dolphin Hitbox: (Doesn't change)